add routing to the page with home, calculator and Quotes pages
