--- src/components/App.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Home from './Home';
-import Quote from './Calculator';
-import Calculator from './Quote';
-
-const App = () => (
-  <BrowserRouter>
-    <Switch>
-      <Route exact path="/" component={Home} />
-      <Route exact path="/calculator" component={Calculator} />
-      <Route exact path="/quote" component={Quote} />
-    </Switch>
-  </BrowserRouter>
-);
-
-export default App;
