+++ src/components/Home.js
@@ -1,5 +1,15 @@
+import Navbar from './Navbar';
+
 const Home = () => (
-  <div>Home</div>
+  <div>
+    <Navbar />
+    <h1>Welcome to my page!</h1>
+    <p>This is my React project that aims to make a simple calculator.</p>
+    <p>
+      It is created using React and lets you make the next operations:
+      addition, substraction, multiplication, division and percentage
+    </p>
+  </div>
 );
 
 export default Home;
