+++ src/components/Quote.js
@@ -1,5 +1,13 @@
+import Navbar from './Navbar';
+
 const Quote = () => (
-  <div>Quote</div>
+  <div>
+    <Navbar />
+    <p>
+      Mathematics is not about numbers, equations, computations,
+      or algorithms: it is about understanding. –William Paul Thurston
+    </p>
+  </div>
 );
 
 export default Quote;
