+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './components/Routes';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(<React.StrictMode><App /></React.StrictMode>, document.getElementById('root'));
