+++ src/style/Button.module.css
@@ -3,5 +3,4 @@
   color: black;
   border: 1px solid black;
   border-radius: 0px;
-  
 }
