+++ src/style/Navbar.module.css
@@ -1,0 +1,7 @@
+.navbar {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 50px;
+}
